+++ waterfall-cad-examples/src/SweepExample.hs
@@ -11,6 +11,7 @@ import qualified Waterfall.Path as Path
 import qualified Waterfall.TwoD.Path2D as Path2D
 import qualified Waterfall.TwoD.Shape as Shape
 import Linear ( V3 (..), (*^), angle, unit, _x, zero)
+import Data.Maybe (fromMaybe)
 
 sweepExample :: Solid
 sweepExample =