+++ waterfall-cad/src/Waterfall/Error.hs
@@ -1,0 +1,10 @@
+module Waterfall.Error
+( WaterfallError (..)
+) where
+
+import Data.Data (Typeable)
+import OpenCascade.Internal.Exception (OpenCascadeException)
+
+newtype WaterfallError = WaterfallError 
+    { rawOpenCascadeError :: OpenCascadeException
+    } deriving (Eq, Ord, Typeable, Show)