+++ waterfall-cad/src/Waterfall/Fillet.hs
@@ -9,17 +9,24 @@ module Waterfall.Fillet
   roundFillet
 , roundConditionalFillet
 , roundIndexedConditionalFillet
+, tryRoundFillet
+, tryRoundConditionalFillet
+, tryRoundIndexedConditionalFillet
 -- * Chamfers
 -- | Adds flat faces at a constant angle to the two faces either side of an edge.
 , chamfer
 , conditionalChamfer
 , indexedConditionalChamfer
+, tryChamfer
+, tryConditionalChamfer
+, tryIndexedConditionalChamfer
 -- * Utility Methods
 , whenNearlyEqual
 ) where
 
-import Waterfall.Internal.Solid (Solid (..), acquireSolid, solidFromAcquire)
+import Waterfall.Internal.Solid (Solid (..), acquireSolid, solidFromAcquireWithCatch)
 import Waterfall.Internal.Edges (edgeEndpoints)
+import Waterfall.Error (WaterfallError)
 import qualified OpenCascade.BRepFilletAPI.MakeFillet as MakeFillet
 import qualified OpenCascade.BRepFilletAPI.MakeChamfer as MakeChamfer
 import qualified OpenCascade.BRepBuilderAPI.MakeShape as MakeShape
@@ -32,8 +39,10 @@ import Control.Monad (when)
 import Control.Monad.IO.Class (liftIO)
 import OpenCascade.Inheritance (upcast, unsafeDowncast)
 import Linear.V3 (V3 (..))
-import Linear.Epsilon 
+import Linear.Epsilon (Epsilon, nearZero)
 import Control.Lens (Lens', (^.))
+import Data.Either (fromRight)
+
 addEdges :: (Integer -> (V3 Double, V3 Double) -> Maybe Double) -> (Double -> Ptr TopoDS.Edge -> IO ()) -> Ptr Explorer.Explorer -> IO ()
 addEdges radiusFn action explorer = go [] 0
     where go visited i = do
@@ -53,6 +62,21 @@ addEdges radiusFn action explorer = go [] 0
                         Explorer.next explorer
                         go (hash:visited) (i + 1) 
 
+
+-- | Version of `roundIndexedConditionalFillet` that returns an `Either` on failure
+tryRoundIndexedConditionalFillet
+    :: (Integer -> (V3 Double, V3 Double) -> Maybe Double)
+    -> Solid
+    -> Either WaterfallError Solid
+tryRoundIndexedConditionalFillet radiusFunction solid = solidFromAcquireWithCatch $ do
+    s <- acquireSolid solid
+    builder <- MakeFillet.fromShape s
+
+    explorer <- Explorer.new s ShapeEnum.Edge
+    liftIO $ addEdges radiusFunction (MakeFillet.addEdgeWithRadius builder) explorer
+
+    MakeShape.shape (upcast builder)
+
 -- | Add rounds with the given radius to each edge of a solid, conditional on the endpoints of the edge, and the index of the edge.
 -- 
 -- This can be used to selectively round\/fillet a `Solid`.
@@ -62,15 +86,19 @@ addEdges radiusFn action explorer = go [] 0
 -- there's no way to select either the curved or the flat edge of the semicircle based on just the endpoints.
 --
 -- Being able to selectively round\/fillet based on edge index is an \"easy\" way to round\/fillet these shapes. 
-roundIndexedConditionalFillet :: (Integer -> (V3 Double, V3 Double) -> Maybe Double) -> Solid -> Solid
-roundIndexedConditionalFillet radiusFunction solid = solidFromAcquire $ do
-    s <- acquireSolid solid
-    builder <- MakeFillet.fromShape s
+roundIndexedConditionalFillet
+    :: (Integer -> (V3 Double, V3 Double) -> Maybe Double)
+    -> Solid
+    -> Solid
+roundIndexedConditionalFillet radiusFunction solid = fromRight mempty $ tryRoundIndexedConditionalFillet radiusFunction solid
 
-    explorer <- Explorer.new s ShapeEnum.Edge
-    liftIO $ addEdges radiusFunction (MakeFillet.addEdgeWithRadius builder) explorer
 
-    MakeShape.shape (upcast builder)
+-- | Version of `roundConditionalFillet` that returns an `Either` on failure
+tryRoundConditionalFillet 
+    :: ((V3 Double, V3 Double) -> Maybe Double)
+    -> Solid 
+    -> Either WaterfallError Solid
+tryRoundConditionalFillet f = tryRoundIndexedConditionalFillet (const f)
 
 -- | Add rounds with the given radius to each edge of a solid, conditional on the endpoints of the edge.
 -- 
@@ -85,6 +113,25 @@ roundFillet :: Double -> Solid -> Solid
 roundFillet r = roundConditionalFillet (const . pure $ r)
 
 
+-- | Version of `roundFillet` that returns an `Either` on failure
+tryRoundFillet :: Double -> Solid -> Either WaterfallError Solid
+tryRoundFillet r = tryRoundConditionalFillet (const . pure $ r)
+
+
+-- | Version of `indexedConditionalChamfer` that returns an `Either` on failure
+tryIndexedConditionalChamfer 
+    :: (Integer -> (V3 Double, V3 Double) -> Maybe Double)
+    -> Solid 
+    -> Either WaterfallError Solid
+tryIndexedConditionalChamfer radiusFunction solid = solidFromAcquireWithCatch $ do
+    s <- acquireSolid solid
+    builder <- MakeChamfer.fromShape s
+
+    explorer <- Explorer.new s ShapeEnum.Edge
+    liftIO $ addEdges radiusFunction (MakeChamfer.addEdgeWithDistance builder) explorer
+
+    MakeShape.shape (upcast builder)
+
 -- | Add chamfers of the given size to each edge of a solid, conditional on the endpoints of the edge, and the index of the edge.
 -- 
 -- This can be used to selectively chamfer a `Solid`.
@@ -94,21 +141,29 @@ roundFillet r = roundConditionalFillet (const . pure $ r)
 -- there's no way to select either the curved or the flat edge of the semicircle based on just the endpoints.
 --
 -- Being able to selectively chamfer based on edge index is an \"easy\" way to chamfer these shapes. 
-indexedConditionalChamfer :: (Integer -> (V3 Double, V3 Double) -> Maybe Double) -> Solid -> Solid
-indexedConditionalChamfer radiusFunction solid = solidFromAcquire $ do
-    s <- acquireSolid solid
-    builder <- MakeChamfer.fromShape s
+indexedConditionalChamfer
+    :: (Integer -> (V3 Double, V3 Double) -> Maybe Double)
+    -> Solid 
+    -> Solid
+indexedConditionalChamfer radiusFunction solid =
+    fromRight mempty $ tryIndexedConditionalChamfer radiusFunction solid
 
-    explorer <- Explorer.new s ShapeEnum.Edge
-    liftIO $ addEdges radiusFunction (MakeChamfer.addEdgeWithDistance builder) explorer
-
-    MakeShape.shape (upcast builder)
+-- | Version of `conditionalChamfer` that returns an `Either` on failure
+tryConditionalChamfer 
+    :: ((V3 Double, V3 Double) -> Maybe Double) 
+    -> Solid
+    -> Either WaterfallError Solid
+tryConditionalChamfer f = tryIndexedConditionalChamfer (const f)
 
 -- | Add chamfers with the given size to each edge of a solid, conditional on the endpoints of the edge.
 -- 
 -- This can be used to selectively chamfer a `Solid`.
 conditionalChamfer :: ((V3 Double, V3 Double) -> Maybe Double) -> Solid -> Solid
 conditionalChamfer f = indexedConditionalChamfer (const f)
+
+-- | Version of `chamfer` that returns an `Either` on failure
+tryChamfer :: Double -> Solid -> Either WaterfallError Solid
+tryChamfer d = tryConditionalChamfer (const . pure $ d)
 
 -- | Add a chamfer with a given size to every edge of a solid
 --