+++ waterfall-cad/src/Waterfall/Internal/Solid.hs
@@ -4,6 +4,7 @@ module Waterfall.Internal.Solid
 ( Solid (..)
 , acquireSolid
 , solidFromAcquire
+, solidFromAcquireWithCatch
 , union3D
 , difference3D
 , intersection3D
@@ -28,9 +29,10 @@ import qualified OpenCascade.BOPAlgo.Operation as BOPAlgo.Operation
 import qualified OpenCascade.BOPAlgo.BOP as BOPAlgo.BOP
 import qualified OpenCascade.BOPAlgo.Builder as BOPAlgo.Builder
 import OpenCascade.Inheritance (upcast)
-import Waterfall.Internal.Finalizers (toAcquire, unsafeFromAcquire)
+import Waterfall.Internal.Finalizers (toAcquire, unsafeFromAcquire, unsafeFromAcquireWithCatch)
 import qualified OpenCascade.BOPAlgo.Builder as BOPAlgo
 import Data.Foldable (traverse_)
+import Waterfall.Error (WaterfallError)
 
 -- | The Boundary Representation of a solid object.
 --
@@ -50,6 +52,8 @@ acquireSolid (Solid ptr) = toAcquire ptr
 solidFromAcquire :: Acquire (Ptr TopoDS.Shape.Shape) -> Solid
 solidFromAcquire = Solid . unsafeFromAcquire
 
+solidFromAcquireWithCatch :: Acquire (Ptr TopoDS.Shape.Shape) -> Either WaterfallError Solid
+solidFromAcquireWithCatch = fmap Solid . unsafeFromAcquireWithCatch
 
 -- | print debug information about a Solid when it's evaluated 
 -- exposes the properties of the underlying OpenCacade.TopoDS.Shape